+++ packages/apps/file-explorer/src/main.ts
@@ -1,8 +1,9 @@
-import { App } from "@prozilla-os/core";
+import { App, AppsConfig } from "@prozilla-os/core";
 import { FileExplorer, FileExplorerProps } from "./components/FileExplorer";
 
 const fileExplorer = new App<FileExplorerProps>("File Explorer", "file-explorer", FileExplorer)
-	.setIconUrl("https://os.prozilla.dev/assets/apps/icons/file-explorer.svg");
+	.setIconUrl("https://os.prozilla.dev/assets/apps/icons/file-explorer.svg")
+	.setRole(AppsConfig.APP_ROLES.FileExplorer);
 
 export { fileExplorer };
 export { FileSelectorMode } from "./types/utils";