+++ packages/apps/terminal/src/main.ts
@@ -1,7 +1,8 @@
-import { App } from "@prozilla-os/core";
+import { App, AppsConfig } from "@prozilla-os/core";
 import { Terminal, TerminalProps } from "./components/Terminal";
 
 const terminal = new App<TerminalProps>("Terminal", "terminal", Terminal)
-	.setIconUrl("https://os.prozilla.dev/assets/apps/icons/terminal.svg");
+	.setIconUrl("https://os.prozilla.dev/assets/apps/icons/terminal.svg")
+	.setRole(AppsConfig.APP_ROLES.Terminal);
 
 export { terminal };