+++ packages/apps/text-editor/src/components/overrides/MarkdownLink.tsx
@@ -2,7 +2,7 @@ import { MouseEventHandler, ReactNode, useMemo } from "react";
 import { faClipboard, faExternalLink } from "@fortawesome/free-solid-svg-icons";
 import { MarkdownProps } from "../TextEditor";
 import styles from "../TextEditor.module.css";
-import { Actions, ClickAction, copyToClipboard, DIALOG_CONTENT_TYPES, DialogBox, ModalProps, removeUrlProtocol, TextDisplay, useContextMenu, useWindowedModal, Vector2 } from "@prozilla-os/core";
+import { Actions, ClickAction, copyToClipboard, DialogBox, ModalProps, ModalsConfig, removeUrlProtocol, TextDisplay, useContextMenu, useWindowedModal, Vector2 } from "@prozilla-os/core";
 import { sanitizeProps } from "../../core/_utils/sanitizeProps";
 
 interface MarkdownLinkProps extends MarkdownProps {
@@ -32,7 +32,7 @@ export function MarkdownLink({ href, children, windowsManager, currentFile, setC
 					Modal: (props: ModalProps) =>
 						<DialogBox {...props}>
 							<p>Target not found: "{href}"</p>
-							<button data-type={DIALOG_CONTENT_TYPES.closeButton}>Ok</button>
+							<button data-type={ModalsConfig.DIALOG_CONTENT_TYPES.CloseButton}>Ok</button>
 						</DialogBox>
 				});
 			}