+++ packages/core/src/constants/modals.const.ts
@@ -1,3 +1,0 @@
-export const DIALOG_CONTENT_TYPES = {
-	closeButton: 0
-};