+++ packages/apps/file-explorer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prozilla-os/file-explorer
 
+## 1.0.10
+
+### Patch Changes
+
+- Added application roles
+- Updated dependencies
+  - @prozilla-os/core@1.0.13
+
 ## 1.0.9
 
 ### Patch Changes